% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{diag_block_matrix}
\alias{diag_block_matrix}
\title{diag_block_matrix}
\usage{
diag_block_matrix(blocks, coeff)
}
\arguments{
\item{blocks}{a vector containing the sizes of the blocks}

\item{coeff}{a vector containing the coefficient corresponding to each block, the coefficients must be between 0 and 1}
}
\value{
a diagonal block-matrix containing the specified coefficients
}
\description{
This function enables the user to generate a diagonal block-matrix with homogeneous blocks
}
\examples{
diag_block_matrix(c(3,4,5),c(0.3,0.4,0.8))

}
