% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{gauss_gen}
\alias{gauss_gen}
\title{gauss_gen}
\usage{
gauss_gen(R, n)
}
\arguments{
\item{R}{a correlation matrix of size dxd}

\item{n}{the number of observations}
}
\value{
a nxd data frame containing n observations of the d variables
}
\description{
This function enables the user to generate gaussian vectors with correlation matrix R
}
\examples{
M <- diag_block_matrix(c(3,4,5),c(0.7,0.8,0.2))
gauss_gen(M,20)

}
