% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\docType{data}
\name{icgc_data}
\alias{icgc_data}
\title{ICGC dataset}
\format{
A dataframe of 15 variables and 250 observations containing the following:
\describe{
  \item{ACACA, AKT1S1, ANLN,ANXA1,AR}{RNA counts (discrete)}
  \item{ACACA_P, AKT1S1_P, ANLN_P,ANXA_P,AR_P}{protein expression measurements (discrete) }
  \item{MU5219,MU4468,MU7870,MU4842,MU6962}{5 mutations (binary)}
}
}
\usage{
icgc_data
}
\description{
Dataset containing RNA counts, protein expression and mutations measured on breast cancer tumors.
}
\keyword{datasets}
